tests/wpt/tests/html/canvas/element/text/2d.text.measure.width.basic.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.text.measure.width.basic</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<style>
@font-face {
  font-family: CanvasTest;
  src: url("/fonts/CanvasTest.ttf");
}
</style>
<body class="show_output">

<h1>2d.text.measure.width.basic</h1>
<p class="desc">The width of character is same as font used</p>


<span style="font-family: CanvasTest; position: absolute; visibility: hidden">A</span>
<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  await document.fonts.ready;
  ctx.font = '50px CanvasTest';
  _assertSame(ctx.measureText('A').width, 50, "ctx.measureText('A').width", "50");
  _assertSame(ctx.measureText('AA').width, 100, "ctx.measureText('AA').width", "100");
  _assertSame(ctx.measureText('ABCD').width, 200, "ctx.measureText('ABCD').width", "200");

  ctx.font = '100px CanvasTest';
  _assertSame(ctx.measureText('A').width, 100, "ctx.measureText('A').width", "100");

}, "The width of character is same as font used");
</script>
